:star2:  Project aggregate GF commands
