+++ src/modules/project/useCases/uploadGF.spec.ts
@@ -68,7 +68,7 @@ describe('uploadGF use-case', () => {
 
     it('should add the GF', () => {
       const fakeFile = fileRepo.save.mock.calls[0][0]
-      expect(fakeProject.addGarantiesFinancieres).toHaveBeenCalledWith(
+      expect(fakeProject.uploadGarantiesFinancieres).toHaveBeenCalledWith(
         gfDate,
         fakeFile.id.toString(),
         user
